Use internal types, fix breakages
