+++ libc/hdr/types/CMakeLists.txt
@@ -31,6 +31,14 @@ add_proxy_header_library(
 )
 
 add_proxy_header_library(
+  struct_f_owner_ex
+  HDRS
+    struct_f_owner_ex.h
+  FULL_BUILD_DEPENDS
+    libc.include.llvm-libc-types.struct_f_owner_ex
+)
+
+add_proxy_header_library(
   struct_timespec
   HDRS
     struct_timespec.h
